修复Payload数据类型错误的BUG。

+++ FirmwareDownloaderHelper/Package.cs
@@ -5,7 +5,7 @@ namespace FirmwareDownloaderHelper
 {
     public class Package
     {
-        protected uint PayloadLength;
+        protected ushort PayloadLength;
 
         protected ushort BasicLength;
